+++ LMeter/ACT/ACTEvent.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using Newtonsoft.Json;
 using System;
 using LMeter.Helpers;
+using System.Text.RegularExpressions;
 
 namespace LMeter.ACT
 {
@@ -38,23 +39,23 @@ namespace LMeter.ACT
 
     public class Encounter
     {
+        [JsonIgnore]
+        private static readonly Random _rand = new Random();
+
+        [JsonIgnore]
+        private static readonly Regex _regex = new Regex(@"\[(\w*)(-k)?\.?(\d+)?\]", RegexOptions.Compiled);
+        
+        [JsonIgnore]
+        private static readonly Dictionary<string, PropertyInfo> _properties = typeof(Encounter).GetProperties().ToDictionary((x) => x.Name.ToLower());
+
         public static string[] GetTags()
         {
             return typeof(Encounter).GetProperties().Select(x => $"[{x.Name.ToLower()}]").ToArray();
         }
 
-        public string GetFormattedString(string format)
+        public string GetFormattedString(string format, string numberFormat)
         {
-            foreach (PropertyInfo prop in this.GetType().GetProperties())
-            {
-                string? value = prop.GetValue(this)?.ToString();
-                if (value is not null)
-                {
-                    format = format.Replace($"[{prop.Name.ToLower()}]", value);
-                }
-            }
-
-            return format;
+            return _regex.Replace(format, new TextTagFormatter(this, numberFormat, _properties).Evaluate);
         }
 
         [JsonProperty("title")]
@@ -67,37 +68,45 @@ namespace LMeter.ACT
         private string _duration { get; set; } = string.Empty;
         
         [JsonProperty("encdps")]
-        public string Dps { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? Dps { get; private set; }
 
         [JsonProperty("damage")]
-        public string DamageTotal { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? DamageTotal { get; private set; }
         
         [JsonProperty("enchps")]
-        public string Hps { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? Hps { get; private set; }
 
         [JsonProperty("healed")]
-        public string HealingTotal { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? HealingTotal { get; private set; }
 
         [JsonProperty("damagetaken")]
-        public string DamageTaken { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? DamageTaken { get; private set; }
 
         [JsonProperty("deaths")]
-        public string Deaths { get; private set; } = string.Empty;
+        public string? Deaths { get; private set; }
 
         [JsonProperty("kills")]
-        public string Kills { get; private set; } = string.Empty;
+        public string? Kills { get; private set; }
 
         public static Encounter GetTestData()
         {
+            float damage = _rand.Next(212345 * 8);
+            float healing = _rand.Next(41234 * 8);
+
             return new Encounter()
             {
-                Duration = "00:15",
+                Duration = "00:30",
                 Title = "Preview",
-                Dps = "69420",
-                Hps = "42069",
+                Dps = new LazyFloat(damage / 30),
+                Hps = new LazyFloat(healing / 30),
                 Deaths = "0",
-                DamageTotal = "69420",
-                HealingTotal = "42069"
+                DamageTotal = new LazyFloat(damage),
+                HealingTotal = new LazyFloat(healing)
             };
         }
     }
@@ -105,25 +114,22 @@ namespace LMeter.ACT
     public class Combatant
     {
         [JsonIgnore]
-        private static Random _rand = new Random();
+        private static readonly Random _rand = new Random();
+
+        [JsonIgnore]
+        private static readonly Regex _regex = new Regex(@"\[(\w*)(-k)?\.?(\d+)?\]", RegexOptions.Compiled);
+
+        [JsonIgnore]
+        private static readonly Dictionary<string, PropertyInfo> _properties = typeof(Combatant).GetProperties().ToDictionary((x) => x.Name.ToLower());
 
         public static string[] GetTags()
         {
             return typeof(Combatant).GetProperties().Select(x => $"[{x.Name.ToLower()}]").ToArray();
         }
 
-        public string GetFormattedString(string format)
+        public string GetFormattedString(string format, string numberFormat)
         {
-            foreach (PropertyInfo prop in this.GetType().GetProperties())
-            {
-                string? value = prop.GetValue(this)?.ToString();
-                if (value is not null)
-                {
-                    format = format.Replace($"[{prop.Name.ToLower()}]", value);
-                }
-            }
-
-            return format;
+            return _regex.Replace(format, new TextTagFormatter(this, numberFormat, _properties).Evaluate);
         }
 
         [JsonProperty("name")]
@@ -136,13 +142,16 @@ namespace LMeter.ACT
         public string Duration { get; private set; } = string.Empty;
         
         [JsonProperty("encdps")]
-        public string EncDps { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? EncDps { get; private set; }
 
         [JsonProperty("dps")]
-        public string Dps { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? Dps { get; private set; }
 
         [JsonProperty("damage")]
-        public string DamageTotal { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? DamageTotal { get; private set; }
 
         [JsonProperty("damage%")]
         public string DamagePct { get; private set; } = string.Empty;
@@ -157,19 +166,23 @@ namespace LMeter.ACT
         public string CritDirectHitPct { get; private set; } = string.Empty;
         
         [JsonProperty("enchps")]
-        public string EncHps { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? EncHps { get; private set; }
         
         [JsonProperty("hps")]
-        public string Hps { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? Hps { get; private set; }
 
         [JsonProperty("healed")]
-        public string HealingTotal { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? HealingTotal { get; private set; }
 
         [JsonProperty("healed%")]
         public string HealingPct { get; private set; } = string.Empty;
 
         [JsonProperty("damagetaken")]
-        public string DamageTaken { get; private set; } = string.Empty;
+        [JsonConverter(typeof(LazyFloatConverter))]
+        public LazyFloat? DamageTaken { get; private set; }
 
         [JsonProperty("deaths")]
         public string Deaths { get; private set; } = string.Empty;
@@ -202,26 +215,26 @@ namespace LMeter.ACT
 
         private static Combatant GetCombatant(params string[] jobs)
         {
-            int damage = _rand.Next(200000);
-            int healing = _rand.Next(50000);
+            int damage = _rand.Next(212345);
+            int healing = _rand.Next(41234);
 
             return new Combatant()
             {
                 Name = "Fake Name",
-                Duration = "00:15",
+                Duration = "00:30",
                 Job = jobs.Select(x => x.ToString()).ElementAt(_rand.Next(jobs.Length)),
-                DamageTotal = damage.ToString(),
-                Dps = (damage / 15).ToString(),
-                EncDps = (damage / 15).ToString(),
-                HealingTotal = healing.ToString(),
-                Hps = (healing / 15).ToString(),
-                EncHps = (healing / 15).ToString(),
+                DamageTotal = new LazyFloat(damage.ToString()),
+                Dps = new LazyFloat((damage / 30).ToString()),
+                EncDps = new LazyFloat((damage / 30).ToString()),
+                HealingTotal = new LazyFloat(healing.ToString()),
+                Hps = new LazyFloat((healing / 30).ToString()),
+                EncHps = new LazyFloat((healing / 30).ToString()),
                 DamagePct = "100%",
                 HealingPct = "100%",
                 CritHitPct = "20%",
                 DirectHitPct = "25%",
                 CritDirectHitPct = "5%",
-                DamageTaken = (damage / 20).ToString(),
+                DamageTaken = new LazyFloat((damage / 20).ToString()),
                 Deaths = _rand.Next(2).ToString(),
                 MaxHit = "Full Thrust-42069"
             };