+++ LMeter/Config/VisibilityConfig.cs
@@ -6,6 +6,7 @@ using Newtonsoft.Json;
 using LMeter.Helpers;
 using System.Linq;
 using System.Collections.Generic;
+using LMeter.ACT;
 
 namespace LMeter.Config
 {
@@ -23,6 +24,7 @@ namespace LMeter.Config
         public bool HideOutsideDuty = false;
         public bool HideWhilePerforming = false;
         public bool HideInGoldenSaucer = false;
+        public bool HideIfNotConnected = false;
 
         public JobType ShowForJobTypes = JobType.All;
         public string CustomJobString = string.Empty;
@@ -60,6 +62,11 @@ namespace LMeter.Config
                 return false;
             }
 
+            if (this.HideIfNotConnected && ACTClient.Status != ConnectionStatus.Connected)
+            {
+                return false;
+            }
+
             if (this.ShowForJobTypes == JobType.All)
             {
                 return true;
@@ -83,6 +90,7 @@ namespace LMeter.Config
                 ImGui.Checkbox("Hide Outside Duty", ref this.HideOutsideDuty);
                 ImGui.Checkbox("Hide While Performing", ref this.HideWhilePerforming);
                 ImGui.Checkbox("Hide In Golden Saucer", ref this.HideInGoldenSaucer);
+                ImGui.Checkbox("Hide While Not Connected to ACT", ref this.HideIfNotConnected);
                 
                 DrawHelpers.DrawSpacing(1);
                 string[] jobTypeOptions = Enum.GetNames(typeof(JobType));