+++ LMeter/changelog.md
@@ -1,3 +1,11 @@
+# Version 0.1.4.0
+- Added advanced text-tag formatting (kilo-format and decimal-format)
+- Added text command to show/hide Meters (/lm toggle)
+- Added option to hide Meter if ACT is not connected
+- Added option to automatically attempt to reconnect to ACT
+- Added option to add gaps between bars
+- Fixed various bugs
+
 # Version 0.1.3.1
 - Make auto-end disabled by default